Restructure list modes to support data sources
We want to support pulling list data from multiple data sources as well
as pushing list data to external systems.
This requires storing multiple sources which is beyond the current scope
of external lists.
User defined lists should serve this role instead and the old user mode
is renamed to view.
This also requires creating an enum for different sources within a
source.
Examples for Spotify are playlists and albums.

This requires a backwards incompatible schema change so also remove
default favorites and change artists to be a JSON array instead of a
CSV.
This allows for better search queries.